Remove convention mappings
